frontend: Rename Item.react.js to Item.tsx
